Split up folderitem
